Retina Image Bug with OS X Resource Bundle
Added a new step for creating a bundle resource. Without it there is a
critical bug for any retina images.

The COMBINE_HIDPI_IMAGES Build Setting must be removed from the
Resource Bundle. It causes .png and @2x.png images to merge into a
single .tiff file.

The pathForResource method will fail when you look for the standard
.png or @2x.png files because they won't exist.

Added an image to show the setting. Path might need to be updated on
Master.