Skip Install Build Setting Breaks App Archives
Added a new step for "Developign the Framework as a Dependent Project"

Any dependent static library or resource bundle needs to set the Skip
Install Build Setting to Yes. Check all build settings.

Otherwise the bundle will be a Generic Xcode Archive and you won't be
able to share Adhoc, validate, or submit it.